Update PyInstaller call to detect M1 architecture or allow manual override (#497)
* The shebang must be on the first line. Delete blanks and move comments. See SC1128.

https://github.com/koalaman/shellcheck/wiki/SC1128

* In POSIX sh, 'source' in place of '.' is undefined.

https://github.com/koalaman/shellcheck/wiki/SC2039

* build.sh: Update PyInstaller call to detect M1 architecture or allow manual override

This commit updates the build.sh script to include logic for detecting the M1 architecture or allowing a manual override. The SPEC_FILE variable is set accordingly based on the provided argument ("m1" or "regular") or the system's architecture. This ensures that the correct spec file is used when calling PyInstaller from within the virtual environment. Additionally, a comment is added to clarify each step of the process.

Note: The virtual environment is deactivated at the end of the script.

* build.sh: Add architecture detection and print statements

- Added architecture detection for M1 and regular architectures
- Print statements added to indicate the detected architecture and spec file being used

* Update build script command for M1 Mac in manually-run-continue.md

The build script command for building on an M1 Mac has been updated to `sh build.sh m1` instead of `sh buildm1.sh`. This change ensures that the correct build script is used when building for an M1 Mac.